refac: use histget instead of autocmds
The autocmds previously used to handle closing and opening of the minimap
have a confusing order or obtuse functionality which makes producing
the right behavior difficult.

We use the brittle (yet efficient) histget() to check the last command
and determine if it was a 'quit' or a 'buffer delete/wipe', as the
minimap behaves differently for each.  This corrects most bugs
around minimap opening/closing.