GH-107 - Fix search getting messed up by buffer switching
Switching buffers changes which `:his /` index contained the most recent
search (not sure why). This change adds an optional argument to
update_highlight, allowing the caller to specify the context of the
call. We change the index we look at based on the context.

If any more need to be added, we can make a list of contexts.