GH-121 - Fix highlighting and minimap scaling
We were stuck in an in-between scaling strategy. I took out some of the more
complicated "cap under this situation, but not that one" in favor of a minimap
that just scales accurately. I left in a setting for people to use to override
this behavior.

Also found/fixed a bug where it would take two saves to actually re-render the minimap.