Fix buffer switching lag by caching the largest line value.
This prevents a rescan and trades time for theoretical accuracy. It's
possible a long line is added (or taken away) after we open the file,
and we won't handle that as well - scaling won't be as nice and search
highlights won't be accurate. The simple fix if that ever becomes the
case is to close and reopen the file, so I believe this tradeoff is
worth it. The file will have undergone significant changes to enter this
corner case.